Fix lib
