+++ Freenove_4WD_Car_For_ESP32/Emotions.h
@@ -174,5 +174,63 @@ byte save_water_right[][8] =
   0x70, 0x20, 0x7E, 0x22, 0x02, 0x8A, 0xAA, 0x77,
 };
 
+//static emotion -left
+byte static_emotion_left[][8] =
+{
+  0x44, 0x28, 0x10, 0x28, 0x44, 0x00, 0x00, 0x07,
+  0x00, 0x00, 0x00, 0x38, 0x00, 0x00, 0x00, 0x07,
+  0x38, 0x44, 0x54, 0x54, 0x44, 0x38, 0x00, 0x03,
+  0x00, 0x04, 0x04, 0x08, 0x30, 0x00, 0x03, 0x04,
+  0x00, 0x10, 0x28, 0x28, 0x00, 0x00, 0x07, 0x04,
+  0x00, 0x28, 0x10, 0x28, 0x00, 0x00, 0x00, 0x07,
+  0x00, 0x00, 0x38, 0x28, 0x6C, 0x00, 0x04, 0x03,
+  0x03, 0x07, 0x06, 0x00, 0x00, 0x01, 0x00, 0x01,
+  0x01, 0x01, 0x39, 0x29, 0x39, 0x01, 0x00, 0x03,
+  0x00, 0x0C, 0x7C, 0x0C, 0x00, 0x00, 0x03, 0x04,
+  0x00, 0x60, 0x7C, 0x60, 0x00, 0x00, 0x03, 0x04,
+  0x00, 0x00, 0x78, 0x00, 0x60, 0x60, 0x04, 0x03,
+  0x00, 0x38, 0x28, 0x38, 0x00, 0x00, 0x03, 0x00,
+  0x00, 0x00, 0x44, 0x38, 0x00, 0x00, 0x03, 0x02,
+  0x00, 0x38, 0x44, 0x00, 0x38, 0x00, 0x03, 0x00,
+  0x00, 0x71, 0x89, 0xA9, 0x89, 0x71, 0x02, 0x02,
+  0x00, 0x3C, 0x42, 0x5A, 0x5A, 0x42, 0x3C, 0x01,
+  0x38, 0x7C, 0x7C, 0x44, 0x38, 0x00, 0x03, 0x04,
+  0x00, 0x30, 0x48, 0x49, 0x11, 0x20, 0x03, 0x01,
+  0x1C, 0x22, 0x41, 0x4C, 0x20, 0x11, 0x08, 0x07,
+  0x38, 0x44, 0x42, 0x12, 0x04, 0x18, 0x03, 0x02
+};
+
+//static emotion -right
+byte static_emotion_right[][8] =
+{
+  0x22, 0x14, 0x08, 0x14, 0x22, 0x00, 0x00, 0xE0,
+  0x00, 0x1C, 0x1C, 0x1C, 0x00, 0x00, 0x20, 0xE0,
+  0x1C, 0x22, 0x2A, 0x2A, 0x22, 0x1C, 0x00, 0xC0,
+  0x00, 0x20, 0x20, 0x10, 0x0C, 0x00, 0xC0, 0x20,
+  0x00, 0x08, 0x14, 0x14, 0x00, 0x00, 0xE0, 0x20,
+  0x00, 0x1C, 0x14, 0x1C, 0x00, 0x20, 0x20, 0xC0,
+  0x00, 0x00, 0x1C, 0x14, 0x36, 0x00, 0x20, 0xC0,
+  0xC0, 0xE0, 0x60, 0x60, 0xC0, 0x80, 0x00, 0x80,
+  0x80, 0x80, 0x9C, 0x94, 0x9C, 0x80, 0x00, 0xC0,
+  0x00, 0x30, 0x3E, 0x30, 0x00, 0x00, 0xC0, 0x20,
+  0x00, 0x06, 0x3E, 0x06, 0x00, 0x00, 0xC0, 0x20,
+  0x00, 0x00, 0x1E, 0x00, 0x06, 0x06, 0x20, 0xC0,
+  0x00, 0x1C, 0x14, 0x1C, 0x00, 0x40, 0xC0, 0x00,
+  0x00, 0x00, 0x22, 0x1C, 0x00, 0x00, 0xC0, 0x40,
+  0x00, 0x1C, 0x22, 0x00, 0x1C, 0x00, 0xC0, 0x00,
+  0x00, 0x8E, 0x91, 0x95, 0x91, 0x8E, 0x40, 0x40,
+  0x00, 0x3C, 0x42, 0x5A, 0x5A, 0x42, 0x3C, 0x80,
+  0x1C, 0x3E, 0x3E, 0x22, 0x1C, 0x00, 0xC0, 0x20,
+  0x00, 0x0C, 0x12, 0x92, 0x88, 0x04, 0xC0, 0x80,
+  0x38, 0x44, 0x82, 0x32, 0x04, 0x88, 0x10, 0xE0,
+  0x1C, 0x22, 0x42, 0x48, 0x20, 0x18, 0xC0, 0x40
+};
+
+
+byte clearEmotion[][8] = {
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+};
+
+
 
 #endif //EMOTIONS_H
