+++ Freenove_4WD_Car_For_ESP32/Freenove_4WD_Car_For_ESP32.h
@@ -13,3 +13,10 @@
 #include "Freenove_WiFi.h"
 
 #endif //FREENOVE_4WD_CAR_FOR_ESP32_H
+
+void Light_Car(int mode);
+void Track_Car(int mode);
+
+extern int carFlag;
+void Car_SetMode(int mode);
+void Car_Select(int mode);
