+++ Freenove_4WD_Car_For_ESP32/Freenove_Photosensitive.h
@@ -3,6 +3,9 @@
 
 #define PHOTOSENSITIVE_PIN   33
 
+extern int light_init_value;            //Set the car's initial environment ADC value
+
+
 void Photosensitive_Setup(void);           //Photosensitive initialization
 int Get_Photosensitive(void);              //Gets the photosensitive resistance value
