+++ Sketches/00.0_Servo_90/00.0_Servo_90.ino
@@ -4,11 +4,11 @@
   Auther      : www.freenove.com
   Modification: 2020/12/18
 **********************************************************************/
-#include "Freenove_4WD_Car_For_ESP32.h"
+#include <Freenove_4WD_Car_For_ESP32.h>
 
 void setup()
 {
-  Servo_Setup();    //Servo initialization
+  PCA9685_Setup();    //Servo initialization
 }
 
 void loop()
