--- Sketches/00.0_Servo_90/debug.cfg
@@ -1,14 +1,0 @@
-# SPDX-License-Identifier: GPL-2.0-or-later
-#
-# Example OpenOCD configuration file for ESP32-WROVER-KIT board.
-#
-# For example, OpenOCD can be started for ESP32 debugging on
-#
-#   openocd -f board/esp32-wrover-kit-3.3v.cfg
-#
-
-# Source the JTAG interface configuration file
-source [find interface/ftdi/esp32_devkitj_v1.cfg]
-set ESP32_FLASH_VOLTAGE 3.3
-# Source the ESP32 configuration file
-source [find target/esp32.cfg]
