--- Sketches/02.2_Ultrasonic_Ranging_Car/Freenove_4WD_Car_For_ESP32.h
@@ -1,38 +1,0 @@
-#ifndef _FREENOVE_4WD_CAR_H
-#define _FREENOVE_4WD_CAR_H
-
-#include <Arduino.h>
-#include <PCA9685.h>
-
-#define MOTOR_1_DIRECTION     1 //If the direction is reversed, change 1 to -1
-#define MOTOR_2_DIRECTION     1 //If the direction is reversed, change 1 to -1
-#define MOTOR_3_DIRECTION     1 //If the direction is reversed, change 1 to -1
-#define MOTOR_4_DIRECTION     1 //If the direction is reversed, change 1 to -1
-
-void PCA9685_Close_Com_Address(void);//Close the PCA9685 public address
-
-/////////////////////PCA9685 drive area//////////////////////////////////////
-void PCA9685_Setup(void);              //servo initialization
-void Servo_1_Angle(float angle);     //Set the rotation parameters of servo 1, and the parameters are 0-180 degrees
-void Servo_2_Angle(float angle);     //Set the rotation parameters of servo 2, and the parameters are 0-180 degrees
-void Servo_Sweep(int servo_id, int angle_start, int angle_end);//Servo sweep function;
-void Motor_Move(int m1_speed, int m2_speed, int m3_speed, int m4_speed);//A function to control the car motor
-
-//////////////////////Buzzer drive area////////////////////////////////////
-void Buzzer_Setup(void);                //Buzzer initialization
-void Buzzer_Alert(int beat, int rebeat);//Buzzer alarm function
-
-////////////////////Battery drive area/////////////////////////////////////
-int Get_Battery_Voltage_ADC(void);   //Gets the battery ADC value
-float Get_Battery_Voltage(void);     //Get the battery voltage value
-void Set_Battery_Coefficient(float coefficient);//Set the partial pressure coefficient
-
-////////////////////Photosensitive drive area//////////////////////////////
-void Photosensitive_Setup(void);           //Photosensitive initialization
-int Get_Photosensitive(void);              //Gets the photosensitive resistance value
-
-/////////////////////Ultrasonic drive area/////////////////////////////////
-void Ultrasonic_Setup(void);//Ultrasonic initialization
-float Get_Sonar(void);//Obtain ultrasonic distance data
-
-#endif
