+++ Sketches/03.2_Photosensitive_Car/03.2_Photosensitive_Car.ino
@@ -5,7 +5,7 @@
   Modification: 2020/12/18
 **********************************************************************/
 #include <Arduino.h>
-#include "Freenove_4WD_Car_For_ESP32.h"
+#include <Freenove_4WD_Car_For_ESP32.h>
 
 int photosensitive_sensitivity = 100;// виставте чутливість фоторезистора   
 int photosensitive_init_value = 0;   // початкове значення освітленості
