--- Sketches/03.2_Photosensitive_Car/Array.h
@@ -1,173 +1,0 @@
-//Turn your eyes clockwise
-byte eyeRotate1[][8]  = {
-  0x00, 0x3C, 0x4E, 0x4E, 0x7E, 0x7E, 0x3C, 0x00,
-  0x00, 0x3C, 0x66, 0x66, 0x7E, 0x7E, 0x3C, 0x00,
-  0x00, 0x3C, 0x72, 0x72, 0x7E, 0x7E, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x72, 0x72, 0x7E, 0x3C, 0x00,//
-  0x00, 0x3C, 0x7E, 0x7E, 0x72, 0x72, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x7E, 0x66, 0x66, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x7E, 0x4E, 0x4E, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x4E, 0x4E, 0x7E, 0x3C, 0x00,//
-};
-
-//Turn your eyes counterclockwise
-byte eyeRotate2[][8] = {
-  0x00, 0x3C, 0x72, 0x72, 0x7E, 0x7E, 0x3C, 0x00,
-  0x00, 0x3C, 0x66, 0x66, 0x7E, 0x7E, 0x3C, 0x00,
-  0x00, 0x3C, 0x4E, 0x4E, 0x7E, 0x7E, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x4E, 0x4E, 0x7E, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x7E, 0x4E, 0x4E, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x7E, 0x66, 0x66, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x7E, 0x72, 0x72, 0x3C, 0x00,
-  0x00, 0x3C, 0x7E, 0x72, 0x72, 0x7E, 0x3C, 0x00,
-};
-
-byte eyeBlink[][8] = {
-  0x00, 0x00, 0x3C, 0x66, 0x42, 0x66, 0x3C, 0x00,
-  0x00, 0x00, 0x3C, 0x7E, 0x42, 0x66, 0x3C, 0x00,
-  0x00, 0x00, 0x3C, 0x7E, 0x7E, 0x7E, 0x3C, 0x00,
-  0x00, 0x00, 0x3C, 0x7E, 0x42, 0x66, 0x3C, 0x00,
-  0x00, 0x00, 0x3C, 0x7E, 0x7E, 0x7E, 0x3C, 0x00,
-};
-byte eyeBlink1[][8] = {
-  0x00, 0x18, 0x24, 0x5A, 0x42, 0x3C, 0x00, 0x00,
-  0x00, 0x18, 0x24, 0x42, 0x42, 0x3C, 0x00, 0x00,
-  0x00, 0x18, 0x24, 0x5A, 0x42, 0x3C, 0x00, 0x00,
-  0x00, 0x18, 0x24, 0x42, 0x42, 0x3C, 0x00, 0x00,
-};
-
-byte eyeSmile[][8] = {
-  0x00, 0x00, 0x18, 0x24, 0x42, 0x00, 0x00, 0x00,
-  0x00, 0x18, 0x24, 0x42, 0x00, 0x00, 0x00, 0x00,
-};
-
-byte eyeCry1[][8] = {
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x18, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x18, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x18, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x18,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-};
-
-byte eyeCry2[][8] = {
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x18, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x18, 0x00, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x18, 0x00,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x18,
-  0x00, 0x00, 0x7E, 0x3C, 0x00, 0x00, 0x00, 0x00,
-};
-
-#define  A  15
-byte arrow_up[8][8] =
-{
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, A, A, A, A, 0, 0},
-  {0, A, 0, A, A, 0, A, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-};
-byte arrow_dowm[8][8] =
-{
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-  {0, A, 0, A, A, 0, A, 0},
-  {0, 0, A, A, A, A, 0, 0},
-  {0, 0, 0, A, A, 0, 0, 0},
-};
-byte arrow_left[8][8] =
-{
-  {0, 0, 0, 0, 0, 0, 0, 0},
-  {0, 0, A, 0, 0, 0, 0, 0},
-  {0, A, 0, 0, 0, 0, 0, 0},
-  {A, A, A, A, A, A, A, A},
-  {A, A, A, A, A, A, A, A},
-  {0, A, 0, 0, 0, 0, 0, 0},
-  {0, 0, A, 0, 0, 0, 0, 0},
-  {0, 0, 0, 0, 0, 0, 0, 0},
-};
-byte arrow_right[8][8] =
-{
-  {0, 0, 0, 0, 0, 0, 0, 0},
-  {0, 0, 0, 0, 0, A, 0, 0},
-  {0, 0, 0, 0, 0, 0, A, 0},
-  {A, A, A, A, A, A, A, A},
-  {A, A, A, A, A, A, A, A},
-  {0, 0, 0, 0, 0, 0, A, 0},
-  {0, 0, 0, 0, 0, A, 0, 0},
-  {0, 0, 0, 0, 0, 0, 0, 0},
-};
-byte car_left[][8] =
-{
-  {0, 0, 0, 0, 0, 0, 0, 0},
-  {0, A, A, A, 0, A, A, A},
-  {0, 0, A, 0, 0, 0, A, 0},
-  {A, A, A, A, A, A, A, 0},
-  {A, A, A, A, A, A, A, 0},
-  {0, 0, A, 0, 0, 0, A, 0},
-  {0, A, A, A, 0, A, A, A},
-  {0, 0, 0, 0, 0, 0, 0, 0},
-};
-
-byte car_right[][8] =
-{
-  {0, 0, 0, 0, 0, 0, 0, 0},
-  {A, A, A, 0, A, A, A, 0},
-  {0, A, 0, 0, 0, A, 0, 0},
-  {0, A, A, A, A, A, A, A},
-  {0, A, A, A, A, A, A, A},
-  {0, A, 0, 0, 0, A, 0, 0},
-  {A, A, A, 0, A, A, A, 0},
-  {0, 0, 0, 0, 0, 0, 0, 0},
-};
-
-byte wheel_left[][8] =
-{
-  0x18, 0x10, 0x10, 0x9F, 0xF9, 0x08, 0x08, 0x18,
-  0x20, 0x42, 0x25, 0x18, 0x18, 0xA4, 0x42, 0x04,
-  0x0C, 0x08, 0x88, 0xF8, 0x1F, 0x11, 0x10, 0x30,
-};
-
-byte wheel_right[][8] =
-{
-  0x18, 0x08, 0x08, 0xF9, 0x9F, 0x10, 0x10, 0x18,
-  0x04, 0x42, 0xA4, 0x18, 0x18, 0x25, 0x42, 0x20,
-  0x30, 0x10, 0x11, 0x1F, 0xF8, 0x88, 0x08, 0x0C,
-};
-
-byte I_love_you[][8] =
-{
-  0x00, 0xE0, 0x42, 0x47, 0x47, 0x43, 0xE1, 0x00,
-  0x00, 0x05, 0x45, 0xE5, 0xE2, 0xC2, 0x82, 0x00,
-};
-
-byte save_water_left[][8] =
-{
-  0xFE, 0x82, 0xAA, 0xEE, 0x82, 0xBA, 0xAA, 0xFE, 
-  0xFE, 0x82, 0xAA, 0xEE, 0x82, 0xBA, 0xAA, 0xFE, 
-  0xFE, 0x82, 0xAA, 0xEE, 0x82, 0xBA, 0xAA, 0xFE, 
-  0xFE, 0x82, 0xAA, 0xEE, 0x82, 0xBA, 0xAA, 0xFE,
-};
-
-byte save_water_right[][8] =
-{
-  0x70, 0x20, 0x7E, 0x22, 0x02, 0x8A, 0x8A, 0x77,
-  0x70, 0x20, 0x7E, 0x22, 0x22, 0x8A, 0x8A, 0x77, 
-  0x70, 0x20, 0x7E, 0x22, 0x02, 0xAA, 0x8A, 0x77, 
-  0x70, 0x20, 0x7E, 0x22, 0x02, 0x8A, 0xAA, 0x77,
-};
